mais treinamento

+++ React/gamagit/src/Pages/Repositories/index.js
@@ -3,16 +3,21 @@
 import React, { useEffect, useState } from 'react';
 // abaixo vou importar tudo e associar a variável S
 import * as S from './styled';
+import { useHistory } from 'react-router-dom';
 
 
 export default function Repositories() {
     const [ repositories, setRepositories ] = useState([]);
-    
+    const history = useHistory();
      useEffect(() => {
         let repositoriesName = localStorage.getItem('repositoriesName');
+        if (repositoriesName !== null) {
         repositoriesName = JSON.parse(repositoriesName);
         setRepositories(repositoriesName);
-        localStorage.clear();
+        localStorage.clear(); 
+        } else {
+            history.push('/')
+        }
     }, []);
     return (
         <S.Container>
@@ -24,7 +29,8 @@ export default function Repositories() {
                 )
             } ) }
             </S.List>
-       
+       {/* <a href="/">Voltar</a> o a é uma ancora no html, mas vamos utilizar um recurso do react para não ficar atualizando a página quando clicarmos em voltar */}
+            <S.LinkHome to="/">Voltar</S.LinkHome>
         </S.Container>
     )
 }
